LPD-17215 Move specific fixture and page to tests module folder
We are also naming files following https://github.com/liferay/liferay-frontend-projects/blob/master/guidelines/general/file_names.md
